[upstream] Kinematic perf (erincatto/box2d#970)
Kinematic perf (erincatto/box2d#970)

- b2Shape_SetMaterial -> b2Shape_SetUserMaterial
- Reworked surface material access on chain shapes
- Removed mouse joint. Functionality is now covered by the motor joint.
- Added b2DynamicTree_QueryAll to support debug draw for shapes that
don't collide
- The user material id is now a 64 bits
- Fixed crash when samples app is minimized
- Added kinematic washer sample
- Fix erincatto/box2d#967
- #968 - draw shapes with 0 category bits and draw sleeping contacts
- Kinematic bodies no longer enter graph coloring
- The joint and contact solvers no longer write the state of kinematic
bodies
- Large kinematic bodies touching many dynamic bodies now has better
multithreaded scaling
- Joints between two non-dynamic bodies no longer simulate
- Minor island splitting optimization suggested by Dirk, use the island
id instead of a bool mark
- Joints now have a constraint damping ratio of 2.0 instead of 0.
Intended to reduce overshoot when limits are violated.
- Restored island debug draw

+++ src/Box2D.NET.Samples/Samples/Bodies/Weeble.cs
@@ -79,12 +79,12 @@ public class Weeble : Sample
         m_explosionMagnitude = 8.0f;
     }
 
-    static float FrictionCallback(float frictionA, int materialA, float frictionB, int materialB)
+    static float FrictionCallback(float frictionA, ulong materialA, float frictionB, ulong materialB)
     {
         return 0.1f;
     }
 
-    static float RestitutionCallback(float restitutionA, int materialA, float restitutionB, int materialB)
+    static float RestitutionCallback(float restitutionA, ulong materialA, float restitutionB, ulong materialB)
     {
         return 1.0f;
     }
